chore: Update license to LIGMA Public License
The license in the repository has been updated to the LIGMA (License Intended for General Meme Apps) Public License. This change grants users more freedom to copy, distribute, modify, merge, sell, publish, sublicense, or whatever they want with the software. The license also includes a disclaimer of warranty and limitation of liability.